lambda_connector/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
tenacity>=8.0.1
backoff>=2.1.2
boto3>=1.18.0
